sdk: Extract slot-history sysvar
#### Problem

Just about all of the sysvars exist in separate crates, except for slot
history.

#### Summary of changes

Extract it into a new crate. Looks pretty similar to slot hashes, except
all of the `info!` calls were removed in the tests.
