+++ sdk/slot-history/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "solana-slot-history"
+description = "Types and utilities for the Solana SlotHistory sysvar."
+documentation = "https://docs.rs/solana-slot-history"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+bv = { workspace = true }
+serde = { workspace = true, optional = true }
+serde_derive = { workspace = true, optional = true }
+solana-clock = { workspace = true, default-features = false }
+
+[features]
+serde = ["dep:serde", "dep:serde_derive", "bv/serde"]
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+all-features = true
+rustdoc-args = ["--cfg=docsrs"]
